Сделано отображение значение сигнала биений во всплывающем окне, запуск данного всплывающего окна при нажатии на надпись РЕГИСТРЫ на главном экране.

+++ app/src/main/java/ru/profitsw200/data/data/BeatSignalFrequencyCalculatorRepositoryImpl.kt
@@ -34,10 +34,10 @@ class BeatSignalFrequencyCalculatorRepositoryImpl : BeatSignalFrequencyCalculato
         lfmPeriod: Double,
         delayTime: Double,
         isSymmetricLfm: Boolean
-    ): Int {
+    ): Double {
         return if (isSymmetricLfm)
-                ((lfmPeriod/(delayTime*lfmDeviationFrequency))*MICRO_SECS_FACTOR).toInt()
-        else (((2*lfmPeriod)/(delayTime*lfmDeviationFrequency))*MICRO_SECS_FACTOR).toInt()
+                ((lfmPeriod/(delayTime*lfmDeviationFrequency))*MICRO_SECS_FACTOR)
+        else (((2*lfmPeriod)/(delayTime*lfmDeviationFrequency))*MICRO_SECS_FACTOR)
     }
 
     private fun calculateBeatSignalPeriod(
