+++ app/src/main/java/ru/profitsw200/data/model/BeatSignalParametersModel.kt
@@ -1,6 +1,6 @@
 package ru.profitsw200.data.model
 
 data class BeatSignalParametersModel(
-    val beatSignalFrequencyKHz: Int,
+    val beatSignalFrequencyKHz: Double,
     val beatSignalPeriodMicroSeconds: Int
 )
